Keep peers in sync after SSL handshake failure in gssapi.
In gss_init_sec_context(), if globus_i_gsi_gss_handshake() returns an
error, the code attempts to fetch the output_token (if anyy) twice.
First in gss_init_sec_context() and again in gss_delete_sec_context().
The second fetch clobbers any data from the first fetch, so the client
never sends it.
The fix is to not pass output_token to gss_delete_sec_context() if it
already contains data.