Fix output payload check in gss_init_sec_context() after SSL handshake.
When globus_i_gsi_gss_handshake() indicates that the SSL handshake has
completed successfully, the check for a final output payload to send is
faulty. It checks output_token before the data has been copied into that
buffer. Instead, it needs to look at context->gss_wbio.

This bug puts the client and server out of sync, dooming the connection.