Keep admin script in sync with init script
Even though there is a comment in the admin script that says:
  # Keep in sync with init script
the admin script is not in sync with the init script.
This commit corrects this.

The admin script ignored the GLOBUS_SEG_PIDFMT setting in the
sysconfig file that the init script uses. With the default value they
still did the same, and there is usually no need to change the
default, so it usually worked anyway.