Upload to "hcc-osg-repo.unl.edu"
This is the actual machine that hosts repo.gridcf.org.  We can't SFTP to
that alias because it's a CloudFlare redirect and, according to my
comment, "doesn't properly handle SSH".

The hosting duty for the OSG repos has moved to a different machine so
repo.opensciencegrid.org no longer points to the machine we used to
upload to.  Upload to hcc-osg-repo.unl.edu instead.