Use the existing socket instead of reestablishing the connection
Use BIO_new_ssl and attach the socket to it, instead of using
BIO_new_ssl_connect which creates a new connection.

The old code sometimes failed if, e.g. we have created an IPv6 socket
and the connect BIO tries to create an IPv4 socket and the server
doesn't accept IPv4 connections.

The OpenSSL version in RHEL/CentOS 7 can use an existing IPv6 socket
if it is attached, but its connect BIO is IPv4 only and can not create
new IPv6 sockets.