Add openssl-devel-engine build requirement on Fedora 41+
The ENGINE API is deprecated in openssl since version 3.0. Fedora is
trying to phase it out, and as a first step they have moved the
associated header files into a separete package. For now the ENGINE
API can still be used if this package is added to the build
requriements for the package, as this commit does. Developers are
encouraged to use the PROVIDER API instead.

https://fedoraproject.org/wiki/Changes/OpensslDeprecateEngine